Fixed the lint issue and addressed the review comments
1. Combine the cluster create req and response structures and rename the file to reflect same.
2. Remove unnecessary methods from models.
